requirements.txt
aiohttp==3.9.0
aiofiles==0.6.0
python-dotenv==1.2.1
